CH4/writer_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
fastavro==1.9.7
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1